.github/scripts/validate-hints/requirements.txt
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
pyyaml>=5.0.0
